requirements.txt
amqp==5.0.2
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.3.0
black==20.8b1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==3.1.5
django-braces==1.14.0
django-celery-beat==2.2.0
django-filter==2.4.0
django-sendgrid-v5==0.9.0
django-templated-mail==1.1.1
django-timezone-field==4.1.1
django-tinymce==3.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.8.4
future==0.18.2
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
promise==2.3
prompt-toolkit==3.0.9
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==2.0.0
pylint==2.6.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2020.5
redis==3.5.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.18.0
Rx==1.6.1
sendgrid==6.4.8
singledispatch==3.4.0.3
six==1.15.0
social-auth-core==4.0.3
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
